restaurant-service/requirements.txt
asgiref==3.8.1
Django==4.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
